refactor housing tag implementation to make it more generic in prep for additional tags
